gar/v2: change how git is used to collect changes for patch generation
Use git add -u instead of git add -A to add changes to the index prior
to the commit used to generate the patch in the makepatch target.
This means that only files existing after the original extract/patch
steps are looked at.  It is now safe to run configure, make changes
and then capture the patch...without grabbing things like config.log, etc.