checkpkg: a bugfix for custom-pkgtrans
The 'set -e' option caused the script to fail when cpio returned a non-zero
exit code.  Armoring the code to use cpio ... || true to keep the script
running if that happens.

Also trying more block offsets, which is not really necessary, but doesn't
hurt either.

This function can be still improved by determining the right offset instead of
trial and error.