postgresql-9.0: shared libraries
Shared libraries have been reworked to be installed under /opt/csw/lib.
CSWlibpq has been turned into a backward compatibility package.  Old binaries
have been checked in binary form.

The number of files in the updated CSWlibpq is smaller than in the previous
revision of CSWlibpq, because some of the files from CSWlibpq were not
necessary and they have been removed.  For example:

libpgtypes.so.1 -> libpgtyles.so.1.2
libpgtypes.so.1.1
libpgtypes.so.1.2

In this scenario, libpgtypes.so.1.1 will be never used and can be removed.