checkpkg: Allow to run more than 1 test per class
When writing tests for checkpkg checks, it would be nice to be able to
write more than one test in one class, instead of creating a new class
for every check.  Adding a level of indirection, a function with takes
a callback function as an argument, allows to do that.