pkgdb-web: Return 'null' when package not found
This way, we can distinguish the case of a package not in catalog from a case
of using a wrong URL.

A JSON 'null' translates to Python None, which evaluates as boolean false.
This way we know there's no package.  When a HTTP 404 error is thrown, it
means that the URL used is wrong.