pkgdb-web: A bugfix for error tag display
The problem was that the function returning all error tags was asked for tags
matching arch='all', and no such tags ever exist: even if the package is
architecture-independent, checks are done always specific to an architecture
architecture and are stored as such.

When the package is arch='all', errors for all architectures are displayed.