checkpkg: Fix a bug doubled errors when re-running
It's funny that I once spent about two hours looking for that bug, and
gave up.  This time I simply saw it, there, in the code.

The problem was that the Run() method of the CheckManager2 class was
unconditionally calling a function with was registering checking
functions.  On the second run of checkpkg, all checks were registered
twice, on the third run -- three times, and so on.

The fix is to keep the state, and only register checks once.

This patch re-enables the unit test and fixes the bug.

Signed-off-by: Maciej Blizinski <maciej@opencsw.org>