pkgdb: Bugfix for pkgname->catalogname mapping
For the purpose of /var/sadm/install/contents file indexing, fake
package objects need to be created.  It was possible that two different
pkgnames were mapped to one catalogname, which caused a catalogname
conflict. This patch removes one problem, when the only difference
between two pkgnames is a number.

Patch by Igor Galić.

Signed-off-by: Maciej Bliziński <maciej@opencsw.org>