integrate-catalogs: Discern vers and REV changes
When stabilizing a catalog, we want to push new package revision, but no
version changes. When comparing catalogs, discern between version and revision
upgrades. When generating the shell script, do annotate which one it is in the
comments, to that it's easy to grep through the file and find all the revision
updates.