pkgdb: When hachoir fails, display the exception
When indexing the whole filesystem, hachoir_parser fails on a number of
files. For now, we're catching the exception and moving on. It might
bite us later, but I have no good idea how to deal with it right now.

This change makes the exception message be printed to stdout.