pkgdb-web: Don't show the full data structure
This used to a very useful feature of pkgdb-web.  You could see the full
python data structure with all the metadata in it.  Unfortunately, as of
January 2013, the amount of data is so large that displaying them can take
several minutes. For now, don't retrieve nor display this data structures at
all.

The only remining way is to get the data from the RESTful interface in JSON
format, and process them locally.