pkgdb: Faster catalog generation
Instead of lots of back-and-forth over the REST interface, we now have
a single endpoint which gives us complete catalog information. All we need to
do is to write that to a text file and call it a catalog.

The server side has to do a little more work, for example filter out all
non-CSW packages from dependencies.