cat-gen: rm "catalog" and "descriptions" on empty
When the catalog is empty, all .pkg and .pkg.gz files are unlinked. To
accompany that, let's delete the "catalog" and "description" files.
Therefore, an empty catalog is represented by a missing catalog and
description files.

This is compatible with chkcat behavior which throws an error on a
signed but empty catalog file: no file to check, no error.