cat-gen: Catalog comparison logic fix
It's one small line for the text editor, but has large logic implications for
the generator.

This would have not happened if I could run unit tests.

I could run unit tests if I had the unit test runner: the "go" command.

I could have the "go" command if it built on Solaris.

But it doesn't.