checkpkg: Better error message
This check can fire if you run sparc+intel builds in parallel and have
arch=all packages in your set. When that happens, checkpkg is first asked to
examine a file with content X, but later the same filename contains content B,
because mgar running on the other architecture overwrote it.