cat-gen: Hack the Makefile to compile
Looks like there are differences in how gccgo looks for packages in comparison
to the mainstream go command. I used truss to see where gccgo looks for the
binary object and I symlinked the needed one. This is probably not the best
way to do it, but all the information I could find on stack overflow would
advice me to use the "go" command which we do not have and will not have until
someone ports it.