pkgdb-web: Add application logging
This logging is added to log specifically from the application to a chosen
file. I was worried there would be a problem with multiple threads and
multiple processes, but testing showed that there isn't any logging from
multiple processes into a single file (with our setup).

The log file paths are hardcoded, we need to find a better way of handling
this to make sure that the web apps can be run on private buildfarms too.