checkpkg: Bugfix for disappearing packages
http://lists.opencsw.org/pipermail/maintainers/2014-May/019245.html

Idempotence is hard.

There is a REST URL which allows you to create relational database entries for
a given package. You send a HTTP PUT request, and the package gets registered.
Before that request, the package's metadata only exist as a JSON blob in the
blob table. In order to make any other operations on the package, an entry in
the svr4_file_stats table needs to be created, and this is what this request
does.

What if you send another PUT request at the same URL? REST semantics require
that multiple calls are allowed and that the result is always the same. My
simple-minded solution was to simply nuke all the rows that were created by
this request, and create them anew. The problem is that the function used
deleted too much: it also deleted rows that represent the present of that
package in catalogs. So deleting all the dependent database rows effectively
removed that package from all the catalogs.

The fix is to modify this piece of code to not delete rows that have been
inserted for other causes than just the content of the package.