Fix the usage of MAKEFLAGS so make options are properly passed to sub-Makefile
$(MAKE) MAKEFLAGS= is supposed to prevent make flags to be passed to the sub-make
command according to http://www.gnu.org/software/make/manual/make.html#Options_002fRecursion.
In practice, it doesn't prevent it for the $(MAKE) command called but only for the other sub-make
that could be launched by this command. The correct way to achieve what we want is to use the
following syntax: MAKEFLAGS= $(MAKE).
This allow for instance the "-j" option to be correctly passed to sub-makefiles.