promote-packages: Support for package deletions
This fixes 2 bugs:

When a package goes from present to absent, we only update the timestamp the
first time we see it's absent. If it continues to be absent, let's not touch
the timestamp.

When a package from from absent to present, but was previously in the timing
data structure, we do need to erase the previous entry and create a new one.
This is the !pti.Present condition.