mGAR v2: Handle uncompressed packages
One way I found to create uncompressed packages is to set MKPACKAGE_ARGS to
"--nocompress" in the build recipe. The problem is that the invocation of
checkpkg unconditionally appends ".gz" to the file name, and this fails for
uncompressed packages. This patch tries to look for "--nocompress" in
MKPACKAGE_ARGS, if it's there, it doesn't add .gz to package names when passed
to checkpkg.